requirements.txt
Flask==2.3.3
aiohttp==3.9.1
gunicorn
asyncio
requests==2.31.0
python-dotenv
flask-socketio
